% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinCols.R
\name{BinCols}
\alias{BinCols}
\title{Binary columns function}
\usage{
BinCols(x, increment = 0.01)
}
\arguments{
\item{x}{The column to be binarized}

\item{increment}{The increment for Equally Probable intervals, defaults to 0.01}
}
\description{
This function transforms a continuous target vector into a series of binary target vectors
}
\examples{
BinCols(trainingY, increment = 0.02)

}
\keyword{binary}

