% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{CalcHDI}
\alias{CalcHDI}
\title{Calculate HDI}
\usage{
CalcHDI(sample.points, epdf, cred = 0.9)
}
\arguments{
\item{sample.points}{Vector of target variable values corresponding to pdf}

\item{epdf}{Vector of epdf sample point values}

\item{cred}{Confidence level}
}
\description{
Takes vector of pdf sample points and corresponding target variable values to
calculate the HDI at desired confidence level
}
\keyword{model}

