% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{CheckLengths}
\alias{CheckLengths}
\title{Check list lengths}
\usage{
CheckLengths(model.list, y.vals)
}
\arguments{
\item{model.list}{List of model objects from TrainLEs function}

\item{y.vals}{Vector of y values being passed}
}
\description{
Ensures two objects are the same length
}
\keyword{model}

