% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{CheckModelList}
\alias{CheckModelList}
\title{Check model list object}
\usage{
CheckModelList(model.list)
}
\arguments{
\item{model.list}{List of model objects from TrainLEs function}
}
\description{
Ensures model.list object contains all `train` objects
}
\keyword{model}

