% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{ExtractAccuracy}
\alias{ExtractAccuracy}
\title{Extract model accuracy}
\usage{
ExtractAccuracy(model)
}
\arguments{
\item{model}{The model to extract accuracy from}
}
\description{
Extracts accuracy from best model performance
}
\keyword{model}

