% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{ExtractAccuracySD}
\alias{ExtractAccuracySD}
\title{Extract model accuracy standard deviation}
\usage{
ExtractAccuracySD(model)
}
\arguments{
\item{model}{The model to extract accuracy SD from}
}
\description{
Extracts accuracy standard deviation of resamples from best model performance
}
\keyword{model}

