% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{ExtractBestPreds}
\alias{ExtractBestPreds}
\title{Extract best model predictions}
\usage{
ExtractBestPreds(model)
}
\arguments{
\item{model}{The model to extract predictions from}
}
\description{
Extracts best predictions from each resample from cv data
}
\keyword{model}

