% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{ExtractDownProbs}
\alias{ExtractDownProbs}
\title{Extract probabilities from class estimation predictions}
\usage{
ExtractDownProbs(model)
}
\arguments{
\item{model}{The model to extract probabilities from}
}
\description{
Extracts the probability that instances will be less than the target value
}
\keyword{model}

