% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{ExtractKappa}
\alias{ExtractKappa}
\title{Extract model kappa}
\usage{
ExtractKappa(model)
}
\arguments{
\item{model}{The model to extract kappa from}
}
\description{
Extracts kappa from best model performance
}
\keyword{model}

