% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{ExtractKappaSD}
\alias{ExtractKappaSD}
\title{Extract model kappa standard deviation}
\usage{
ExtractKappaSD(model)
}
\arguments{
\item{model}{The model to extract kappa SD from}
}
\description{
Extracts kappa standard deviation of resamples from best model performance
}
\keyword{model}

