% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractModelInfo.R
\name{ExtractModelInfo}
\alias{ExtractModelInfo}
\title{Extract Model Information}
\usage{
ExtractModelInfo(model.list)
}
\arguments{
\item{model.list}{List of models trained using the TrainLEs function}
}
\description{
This function extracts the performance information and best predictions from a model list
}
\examples{

}
\keyword{model,}
\keyword{predictions,}
\keyword{train}

