% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractModelInfo2.R
\name{ExtractModelInfo2}
\alias{ExtractModelInfo2}
\title{Extract Model Information2}
\usage{
ExtractModelInfo2(model.list)
}
\arguments{
\item{model.list}{List of models trained using the TrainLEs function}
}
\description{
This function extracts the performance information and best predictions from a model list
}
\examples{

}
\keyword{model,}
\keyword{predictions,}
\keyword{train}

