% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{ExtractPerformance}
\alias{ExtractPerformance}
\title{Extract model performance}
\usage{
ExtractPerformance(model)
}
\arguments{
\item{model}{The model to extract performance from}
}
\description{
Extracts all performance data from best model
}
\keyword{model}

