% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotLEs.R
\name{PlotLEs}
\alias{PlotLEs}
\title{Plot Local Expert summary function}
\usage{
PlotLEs(model.info)
}
\arguments{
\item{model.info}{A model information object created using the ExtractModelInfo function}
}
\description{
This plots the average accuracy and kappa values from the resamples of local experts, in addition to their standard deviation
}
\examples{

}
\keyword{plot}

