% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictLEs.R
\name{PredictLEs}
\alias{PredictLEs}
\title{Predict response of new instance using local expert ensemble with stacking}
\usage{
PredictLEs(x, model.list)
}
\arguments{
\item{x}{Attribute vector of new instance to be predicted}

\item{model.list}{Model list object from
\code{\link{TrainLEs}} function}
}
\value{
Returns a vector of probabilies in the [0,1] interval of length equal
to the number of local experts in the model list object.
}
\description{
Generates local expert predictions for a new feature vector \code{x}.
}

