% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictNew.R
\name{PredictNew}
\alias{PredictNew}
\title{Predict response of new instance using local expert ensemble with stacking}
\usage{
PredictNew(x, LE.model.list, stack.model, y.vals, plot = FALSE,
  verbose = FALSE, df = 10, cred, ...)
}
\arguments{
\item{x}{Attribute vector of new instance to be predicted}

\item{LE.model.list}{Model list object from TrainLEs function}

\item{stack.model}{Regression model trained on LE output}

\item{y.vals}{Y-values from BinCols function}

\item{plot}{Display plot of LE predictions and stack estimate}

\item{verbose}{Determines whether prediction information is printed}

\item{df}{Degrees of freedom to pass spline if plotting}

\item{cred}{Credibility interval for HDI in plot}

\item{...}{Additional arguments to pass to the FitInstance function}
}
\description{
Fits a curve to LE predictions, outputs fitted curve point estimates and distribution moments
}
\examples{

}
\keyword{predict}

