% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainLEs.R
\name{TrainLEs}
\alias{TrainLEs}
\title{Train Local Experts}
\usage{
TrainLEs(x, bincols, trControl = NULL, method = "lda", n.folds = 5,
  metric = "Kappa", JIT = FALSE, ...)
}
\arguments{
\item{x}{The feature matrix}

\item{bincols}{Binary target matrix created with 'BinCols()' function}

\item{trControl}{Optional argument to specify a train control object- defaults to cross validated}

\item{method}{Type of learning algorithm used for induction- defaults to lda}

\item{n.folds}{Number of folds if default train control object is used- defaults to 5}

\item{metric}{Optimization metric; can be either "Accuracy" or "Kappa"}

\item{JIT}{Whether or not just-in-time compilation is enabled}

\item{...}{Additional parameters to pass to model training function}
}
\description{
This function trains a list of local expert (level 0) models on a binary response matrix.
}
\examples{

}
\keyword{train}

