% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinCols.R
\name{BinCols}
\alias{BinCols}
\title{Binary columns function}
\usage{
BinCols(y, n = 30, mode = "EW", EW.buffer = 0.01)
}
\arguments{
\item{y}{The column to be binarized}

\item{n}{The number of columns in the resulting binary matrix}

\item{mode}{Defaults to 'EW' intervals, but 'EP' is also available}

\item{EW.buffer}{Defaults to 0.01, only for EW testing; excludes outliers from width}
}
\value{
A matrix of width equal to \code{n} and height equal to \code{length(y)}
}
\description{
This function takes a continuous target vector (ie. the `y` vector in regression)
and generates a matrix of binary target vectors.
}
\details{
The options for \code{mode} are for equally probable bin sizes ('EP') and
equal width bins ('EW') in terms of the values of the target variable. The
parameter \code{n} corresponds to the number of models that will be in the final
local expert ensemble.
}
\examples{
y <- runif(1000, 0, 200)
binary.matrix <- BinCols(y, n = 15, mode = 'EP')

}
\keyword{binary}

