% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractModelInfo.R
\name{ExtractModelInfo}
\alias{ExtractModelInfo}
\title{Extract Model Information}
\usage{
ExtractModelInfo(model.list)
}
\arguments{
\item{model.list}{List of models stored as \code{train} objects}
}
\value{
Returns a list of performance metrics, as well as a matrix
storing the predictions made by the local expert models in the best-tuned
model (if parameters are tuned).
}
\description{
This function extracts performance information and best predictions from
a model list generated by the \code{\link{TrainLEs}} function.
}
\examples{
ExtractModelInfo(neural.net.list)
}

