% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitInstance.R
\name{FitInstance}
\alias{FitInstance}
\title{Fit a CDF to a row of local expert predictions}
\usage{
FitInstance(LE.preds, y.values, granularity = 1000, df = 10, plot = FALSE)
}
\arguments{
\item{LE.preds}{Vector of LE predictions}

\item{y.values}{Y-values from BinCols function}

\item{granularity}{Number of interpolation points in fitting; defaults to 100}

\item{df}{Degrees of freedom in smoothing spline; defaults to 10}

\item{plot}{Plot output; defaults to FALSE}
}
\description{
Fits a curve to LE predictions, outputs fitted curve point estimates and distribution moments
}
\examples{

}
\keyword{smooth}

