% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitInstance.R
\name{FitInstance}
\alias{FitInstance}
\title{Fit a CDF to a row of local expert predictions}
\usage{
FitInstance(row, y.values, granularity = 100, df = 10, mode = "CDF",
  plot = FALSE)
}
\arguments{
\item{row}{Vector of LE predictions}

\item{granularity}{Number of interpolation points in fitting; defaults to 100}

\item{df}{Degrees of freedom in smoothing spline; defaults to 10}

\item{mode}{Choose whether probability output is in CDF or PDF format}

\item{plot}{Plot output; defaults to FALSE}

\item{y.vals}{Y-values from BinCols function}
}
\description{
Fits a curve to LE predictions, outputs fitted curve point estimates and distribution moments
}
\examples{

}
\keyword{smooth}

