% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictLEs.R
\name{PredictLEs}
\alias{PredictLEs}
\title{Predict response of new instance using local expert ensemble with stacking}
\usage{
PredictLEs(x, LE.model.list)
}
\arguments{
\item{x}{Attribute vector of new instance to be predicted}

\item{LE.model.list}{Model list object from TrainLEs function}
}
\description{
Fits a curve to LE predictions, outputs fitted curve point estimates and distribution moments
}
\examples{

}
\keyword{predict}

