% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainLEs.R
\name{TrainLEs}
\alias{TrainLEs}
\title{Train Local Experts}
\usage{
TrainLEs(x, bincols, trControl = NULL, method = "lm", n.repeats = 10, ...)
}
\arguments{
\item{x}{The feature matrix}

\item{bincols}{Binary target matrix created with 'BinCols()' function}

\item{trControl}{Optional argument to specify a train control object- defaults to cross validated}

\item{method}{Type of learning algorithm used for induction- defaults to linear model}

\item{n.repeats}{Number of repeats if default train control object is used- defaults to 10}

\item{...}{Additional parameters to pass to model training function}
}
\description{
This function trains a list of local expert (level 0) models on a binary response matrix.
}
\examples{

}
\keyword{train}

