% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrainLEs.R
\name{TrainLEs}
\alias{TrainLEs}
\title{Train Local Experts}
\usage{
TrainLEs(x, bincols, trControl = NULL, method = "lda", n.folds = 5,
  metric = "Kappa", JIT = FALSE, ...)
}
\arguments{
\item{x}{The feature matrix}

\item{bincols}{Binary target matrix created with 'BinCols()' function}

\item{trControl}{Optional argument to specify a train control object- defaults to cross validated}

\item{method}{Type of learning algorithm used for induction- defaults to lda}

\item{n.folds}{Number of folds if default train control object is used- defaults to 5}

\item{metric}{Optimization metric; can be either "Accuracy" or "Kappa"}

\item{JIT}{Whether or not just-in-time compilation is enabled}

\item{...}{Additional parameters to pass to caret::train}
}
\value{
Returns a list containing \code{train} type objects. List length is
dependent on the number of columns in the matrix supplied as the argument to
\code{bincols}
}
\description{
This function trains a list of local expert models on a binary matrix.
}
\details{
This function calls the \code{\link[caret]{train}} function independently on
each column of the binary matrix supplied to the object supplied to the
\code{bincols} argument, which should have been generated by the
\code{\link{BinCols}} function. The default optimization metric is the
kappa value, but accuracy is also available. For the \code{method} argument,
only classification algorithms can be used (as opposed to regression).
}
\examples{
TrainLEs(x, y.columns, method = 'svmLinear')

}

